Minor connection establish cleanup
A lot of the logic for establishing a connection has been moved
into the InitConnectionTask which is now renamed to TcpIpConnectionEstablisher.

This way we keep the port logic where it is needed and remove the non
relevant parts from the TcpIpConnectionManager.
